Add files via upload
The starter code by Sasank (CTO of Qure.ai, pytorch contributor, and alumnus of IITB) for classifying ants vs. bees: https://pytorch.org/tutorials/beginner/transfer_learning_tutorial.html. The key feature of this code is that it does not train a model from scratch, but uses transfer learning of a ResNet-18 architecture that is pre-trained on a large dataset (ImageNet) and then only fine-tunes it for the problem at hand. The original code is not belong to me, licensed by Sasank.